+++ packages/battery_plus/battery_plus/test/battery_test.dart
@@ -20,6 +20,9 @@ class MockBatteryPlatform
 
   @override
   Stream<BatteryState> get onBatteryStateChanged => controller.stream;
+
+  @override
+  Future<bool> get isInBatterySaveMode => Future.value(true);
 }
 
 void main() {
@@ -34,6 +37,10 @@ void main() {
 
   test('batteryLevel', () async {
     expect(await battery.batteryLevel, 42);
+  });
+
+  test('isInBatterySaveMode', () async {
+    expect(await battery.isInBatterySaveMode, true);
   });
 
   group('battery state', () {