+++ packages/battery_plus/battery_plus_windows/windows/include/battery_plus_windows/system_battery.h
@@ -16,6 +16,7 @@ class SystemBattery {
   SystemBattery();
   ~SystemBattery();
 
+  int GetBatterySaveMode() const;
   int GetLevel() const;
 
   BatteryStatus GetStatus() const;