+++ packages/battery_plus/battery_plus_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.1.0
+
+- Add method channel for power save mode state
 ## 1.0.1
 
 - Improve documentation